Add broken wall near slots visualization
I mostly play this game on mobile and when
near a wall I tend to just randomly tap
the slot as I can't see what there is and
see what the restrained rotation gives me.

However, this strategy misses some cases,
for example if there is a bar between
two "i", if one of the i is on the other
side of the wall, then rip.

Now we can scroll the board on mobile
by taping the border